Adding a trebuchet

+++ Assets/GuardTower.cs
@@ -9,6 +9,6 @@ public class GuardTower : AttackTower
     public override Alliances Enemies => Alliances.Illigons;
     public override float Cooldown => AttackSpeed.Medium;
     public override int Damage => 5;
-    public override float Range => RangeOptions.Medium;
+    public override int Range => RangeOptions.Medium;
     public override VerticalRegion AttackRegion => VerticalRegion.GroundAndAir;
 }
