+++ Assets/Scripts/Buildings/CrystalAccelerator.cs
@@ -5,13 +5,12 @@ public class CrystalAccelerator : AttackTower
 {
     public override float Cooldown => AttackSpeed.Slow;
     public override int Damage => 10;
-    public override float Range => RangeOptions.Short;
+    public override int Range => RangeOptions.Short;
     public override VerticalRegion AttackRegion => VerticalRegion.Ground;
     public override BuildingType Type => BuildingType.CrystalAccelerator;
     public override Alliances Alliance => Alliances.Player;
     public override Alliances Enemies => Alliances.Illigons;
     protected override int ExpectedNumberOfEnemiesHitByEachProjectile => 12;
-
     private ParticleSystem projectileGenerationAnimation;
 
     protected override void Setup()
