+++ Assets/Scripts/Buildings/FlamethrowerTower.cs
@@ -6,7 +6,7 @@ public class FlamethrowerTower : AttackTower
 {
     public override float Cooldown => AttackSpeed.VeryVeryVeryFast;
     public override int Damage => 1;
-    public override float Range => RangeOptions.Short;
+    public override int Range => RangeOptions.Short;
     public override VerticalRegion AttackRegion => VerticalRegion.Ground;
     public override BuildingType Type => BuildingType.FlamethrowerTower;
     public override Alliances Alliance => Alliances.Player;
